Colocando um await no retorno da função findOne

+++ src/models/index.ts
@@ -18,7 +18,7 @@ const create = async (tarefa: ITaskBody): Promise<ITask> => {
 };
 
 const update = async (id: string, task: string, status: string): Promise<ITask | undefined> => {
-  const verifyId = findOne(id);
+  const verifyId = await findOne(id);
   if (!verifyId) return undefined;
   await model.updateOne({ _id: id, task, status });
 
@@ -31,15 +31,14 @@ const update = async (id: string, task: string, status: string): Promise<ITask |
 
 const findOne =  async (id: string): Promise<ITask | undefined> => {
   const find = await model.findOne({ _id: id });
-  console.log('>>>>>>>', find)
   if (!find) return undefined;
 
   return find;
 };
 
 const remove = async (id: string): Promise<ITask | undefined> => {
-  // const verifyId = findOne(id);
-  // if (!verifyId) return undefined;
+  const verifyId = await findOne(id);
+  if (!verifyId) return undefined;
   await model.deleteOne({ _id: id });
 
   return {
